requirements.txt
langchain
langchain-groq
langgraph
langsmith
chromadb
python-dotenv
streamlit
pypdf
pytesseract
pdf2image
Pillow
pydantic
langchain-community
sentence-transformers
langchain-huggingface
graphviz
langchain-google-genai
google-generativeai
streamlit-mermaid